Rename implementation from `scalableCapital` to `baaderBank`
