+++ tests/brokers/__mocks__/baaderbank/buy/saving_plan_vanguard_from_Oskar.json
