+++ tests/brokers/baaderbank.test.js
@@ -1,11 +1,11 @@
 import { findImplementation } from '../../src';
-import * as scalableCapital from '../../src/brokers/scalableCapital';
+import * as baaderBank from '../../src/brokers/baaderBank';
 import {
   allSamples,
   buySamples,
   sellSamples,
   dividendSamples,
-} from './__mocks__/scalablecapital';
+} from './__mocks__/baaderbank';
 
 describe('Broker: scalable.capital', () => {
   let consoleErrorSpy;
@@ -14,7 +14,7 @@ describe('Broker: scalable.capital', () => {
     test('Can the document parsed with scalable.capital', () => {
       allSamples.forEach(samples => {
         expect(
-          samples.some(item => scalableCapital.canParsePage(item, 'pdf'))
+          samples.some(item => baaderBank.canParsePage(item, 'pdf'))
         ).toEqual(true);
       });
     });
@@ -24,14 +24,14 @@ describe('Broker: scalable.capital', () => {
         const implementations = findImplementation(samples, 'pdf');
 
         expect(implementations.length).toEqual(1);
-        expect(implementations[0]).toEqual(scalableCapital);
+        expect(implementations[0]).toEqual(baaderBank);
       });
     });
   });
 
   describe('Validate buys', () => {
     test('Can the market order be parsed from the document', () => {
-      const activities = scalableCapital.parsePages(buySamples[0]).activities;
+      const activities = baaderBank.parsePages(buySamples[0]).activities;
 
       expect(activities.length).toEqual(1);
       expect(activities[0]).toEqual({
@@ -49,7 +49,7 @@ describe('Broker: scalable.capital', () => {
     });
 
     test('Can the saving plan order be parsed from the document - vanguard', () => {
-      const activities = scalableCapital.parsePages(buySamples[1]).activities;
+      const activities = baaderBank.parsePages(buySamples[1]).activities;
 
       expect(activities.length).toEqual(1);
       expect(activities[0]).toEqual({
@@ -67,7 +67,7 @@ describe('Broker: scalable.capital', () => {
     });
 
     test('Can the saving plan order be parsed from the document - comstage', () => {
-      const activities = scalableCapital.parsePages(buySamples[2]).activities;
+      const activities = baaderBank.parsePages(buySamples[2]).activities;
 
       expect(activities.length).toEqual(1);
       expect(activities[0]).toEqual({
@@ -85,7 +85,7 @@ describe('Broker: scalable.capital', () => {
     });
 
     test('Can the limit order be parsed from the document', () => {
-      const activities = scalableCapital.parsePages(buySamples[3]).activities;
+      const activities = baaderBank.parsePages(buySamples[3]).activities;
 
       expect(activities.length).toEqual(1);
       expect(activities[0]).toEqual({
@@ -103,7 +103,7 @@ describe('Broker: scalable.capital', () => {
     });
 
     test('Can the BioNTech order be parsed from the document of Gratisbroker', () => {
-      const activities = scalableCapital.parsePages(buySamples[4]).activities;
+      const activities = baaderBank.parsePages(buySamples[4]).activities;
 
       expect(activities.length).toEqual(1);
       expect(activities[0]).toEqual({
@@ -121,7 +121,7 @@ describe('Broker: scalable.capital', () => {
     });
 
     test('Can a order made from Oskar be parsed from the document', () => {
-      const activities = scalableCapital.parsePages(buySamples[5]).activities;
+      const activities = baaderBank.parsePages(buySamples[5]).activities;
 
       expect(activities.length).toEqual(1);
       expect(activities[0]).toEqual({
@@ -141,7 +141,7 @@ describe('Broker: scalable.capital', () => {
 
   describe('Validate sells', () => {
     test('Can the order be parsed from the document', () => {
-      const activities = scalableCapital.parsePages(sellSamples[0]).activities;
+      const activities = baaderBank.parsePages(sellSamples[0]).activities;
 
       expect(activities.length).toEqual(1);
       expect(activities[0]).toEqual({
@@ -161,8 +161,7 @@ describe('Broker: scalable.capital', () => {
 
   describe('Validate dividends', () => {
     test('Can the etf dividend be parsed from the document', () => {
-      const activities = scalableCapital.parsePages(dividendSamples[0])
-        .activities;
+      const activities = baaderBank.parsePages(dividendSamples[0]).activities;
 
       expect(activities.length).toEqual(1);
       expect(activities[0]).toEqual({
@@ -180,8 +179,7 @@ describe('Broker: scalable.capital', () => {
     });
 
     test('Can the stock dividend in USD with withholding taxes be parsed from the document', () => {
-      const activities = scalableCapital.parsePages(dividendSamples[1])
-        .activities;
+      const activities = baaderBank.parsePages(dividendSamples[1]).activities;
 
       expect(activities.length).toEqual(1);
       expect(activities[0]).toEqual({
@@ -199,8 +197,7 @@ describe('Broker: scalable.capital', () => {
     });
 
     test('Can the stock dividend from Volkswagen with taxes be parsed from the document of Gratisbroker', () => {
-      const activities = scalableCapital.parsePages(dividendSamples[2])
-        .activities;
+      const activities = baaderBank.parsePages(dividendSamples[2]).activities;
 
       expect(activities.length).toEqual(1);
       expect(activities[0]).toEqual({
